Fixing sounds and added Nimbus UI
I realized the sounds needed to be taken care of on the board, which is why I was having trouble. Also added a (in my opinion anyway) better looking UI. The commented out UI is another one you can try out, just comment out the for loop and un-comment the other. Personally, I found Nimbus to be the best.